Move badge component to better package
